Сделали ревью кода. Настроили инициализацию приложения
