+++ internal/app/app.go
@@ -1,26 +1,68 @@
-package app
+package grpcapp
 
 import (
-	grpcapp "UserServiceAuth/internal/app/grpc"
+	"fmt"
 	"log/slog"
-	"time"
+	"net"
+
+	server "UserServiceAuth/internal/router"
+
+	"google.golang.org/grpc"
 )
 
 type App struct {
-	GRPCSrv *grpcapp.App
+	log        *slog.Logger
+	gRPCServer *grpc.Server
+	port       int
 }
 
-func New(
-	log *slog.Logger,
-	grpcPort int,
-	storagePath string,
-	tokenTTL time.Duration,
-) *App {
-
-	grpcApp := grpcapp.New(log, grpcPort)
+// New создает новый экземпляр приложения gRPC сервера.
+func New(log *slog.Logger, port int) *App {
+	gRPCServer := grpc.NewServer()
+	server.RegisterRouter(gRPCServer)
 
 	return &App{
-		GRPCSrv: grpcApp,
+		log:        log,
+		gRPCServer: gRPCServer,
+		port:       port,
+	}
+}
+
+// MustRun запускает gRPC сервер и вызывает panic в случае ошибки.
+func (a *App) MustRun() {
+	if err := a.Run(); err != nil {
+		panic(err)
+	}
+}
+
+// Run запускает gRPC сервер.
+func (a *App) Run() error {
+	const op = "grpcapp.Run"
+
+	log := a.log.With(
+		slog.String("op", op),
+		slog.Int("port", a.port),
+	)
+
+	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", a.port))
+	if err != nil {
+		return fmt.Errorf("%s: %w", op, err)
 	}
 
+	log.Info("gRPC Server is running", slog.String("addr", lis.Addr().String()))
+
+	if err := a.gRPCServer.Serve(lis); err != nil {
+		return fmt.Errorf("%s: %w", op, err)
+	}
+
+	return nil
+}
+
+// Stop останавливает gRPC сервер.
+func (a *App) Stop() {
+	const op = "grpcapp.Stop"
+
+	a.log.With(slog.String("op", op)).Info("stopping gRPC server", slog.Int("port", a.port))
+
+	a.gRPCServer.GracefulStop()
 }
