+++ internal/config/config.go
@@ -8,10 +8,9 @@ import (
 )
 
 type Config struct {
-	Env         string        `yaml:"env" env-default:"local"`
-	StoragePath string        `yaml:"storage_path" env-required:"true"`
-	TokenTTL    time.Duration `yaml:"token_ttl" env-required:"true"`
-	GRPC        GRPCconfig    `yaml:"grpc" env-required:"true"`
+	Env         string     `yaml:"env" env-default:"local"`
+	StoragePath string     `yaml:"storage_path" env-required:"true"`
+	GRPC        GRPCconfig `yaml:"grpc" env-required:"true"`
 }
 
 type GRPCconfig struct {
