+++ internal/router/server.go
@@ -8,21 +8,18 @@ import (
 	"google.golang.org/grpc"
 )
 
-type serverAPI struct {
-	ssov1.UnimplementedGetPublicKeyServer
+func RegisterRouter(gRPC *grpc.Server) {
+	ssov1.RegisterGetPublicKeyServer(gRPC, &serverAPI{})
 }
 
-func Register(gRPC *grpc.Server) {
-	ssov1.RegisterGetPublicKeyServer(gRPC, &serverAPI{})
-
+type serverAPI struct {
+	ssov1.UnimplementedGetPublicKeyServer
 }
 
 func (s *serverAPI) PublicKey(
 	ctx context.Context,
 	req *ssov1.PublicKeyRequest,
 ) (*ssov1.PublicKeyResponse, error) {
-	return &ssov1.PublicKeyResponse{
-		PublicKey: "publickey12731723929381",
-	}, nil
 
+	return &ssov1.PublicKeyResponse{PublicKey: "publickey12731723929381"}, nil
 }
